==> 01_ola_pandas.ipynb <==
## Bibliotéca Pandas

### Principal bibliotéca para manipulação de dados no Python



# import pandas # Importando bibliotéca

import pandas as pd # pd -> Apelidando amigávelmente o pandas para facilitação de uso 

## Por que usamos Pantas?

# Calculando SEM Pandas

idades = [
    32, 38, 30, 30, 31,
    35, 25, 29, 31, 37,
    27, 23, 36, 33, 32
]

media = sum(idades)/len(idades)
print(f'A média de idades é = {media}')

diffs = 0
for i in idades:
    diffs += (i - media) ** 2

variancia = diffs / (len(idades)-1)

print(f'A variância das idades é = {variancia}')

# Calculando COM Pandas

series_idades = pd.Series(idades) # Atribuindo dados na serie

media_series = series_idades.mean()
variancia_series = series_idades.var()

print(f"Média com Series é = {media_series}")
print(f"Variância com Series é = {variancia_series}")

series_idades.describe()
