==> 02_series.ipynb <==
## Series no Pandas

### Características:

- Conjunto de Valores de uma dimenção
- Ordenado - A ordem defini o índices dos elementos
- Homogêneo - Comporta um tipo de dados por elemento
- Comportamento parecido com Dicionário e Listas
- Aceita valores repetidos, inclusive índices

# Importando bibliotécas

import pandas as pd 

# CRIANÇÃO

series_pandas = pd.Series() #Explícta apenas

print(series_pandas, '\n')

idades = [
    32, 38, 30, 30, 31,
    35, 25, 29, 31, 37,
    27, 23, 36, 33, 39
]

series_idades = pd.Series(idades) # Criando e atribuindo dados na serie
series_idades

### Navegando nos dados:

- `series[ i ]` - Navega a partir do índice (label)
- `series.loc[ i ]` - Exatamente igual a `'series[ i ]'`
- `series.iloc[ i ]` - Navega a partir do posição numérica.

# series[ i ] 

navega_lista = idades[0]
print('idade na posição 0:', navega_lista) # Navega na posição

navega_series = series_idades[0]
print('series no índice 0:', navega_series) # Navega no label do índice

# series[-1] - KeyError
try:
    series_idades[-1] # Tentando acessar a última posição pelo índice
except KeyError as e:
    print(f'Não existe um índice com valor "{e}" gerando um KeyError')

# series.iloc[ i ]

navega_series = series_idades.iloc[-1] # Ignora índece | Acessa posição
print('series na posição -1:', navega_series, "\n") # Navega no índice

navega_series = series_idades.iloc[:3] # Slicer
print('series na posição 0 até 2:', "\n", navega_series, "\n")

navega_series = series_idades.iloc[6:0:-1] # Reverse
print('series na com posição 6 á 0 (invertida):', "\n", navega_series, "\n")

navega_series = series_idades.iloc[::2] # Jump
print('series na com posição pulando um:', "\n", navega_series, "\n")

# Alterando labels padrão do index

idades = [
    32, 38, 30, 30, 30,
    35, 25, 29, 31, 37,
    27, 23, 36, 33, 39
]

indexs = [
    'a', 'b', 'c', 'd', 'e',
    'f', 'g', 'h', 'i', 'j',
    'k', 'l', 'm', 'n', 'o'
]

series_idades = pd.Series(idades, index=indexs)

print(series_idades,"\n")

navega_series = series_idades['h']
print('series no índice "h":', navega_series) 

# series.loc[ i ]

navega_series = series_idades.loc['c'] # Exatamente como series_idades['c']
print('series no índice "c":', navega_series) 


### Funções úteis:

# Atributos

print("Exibindo todos as labels dos índices com atributo '.index':", '\n', series_idades.index, '\n')

print("Exibindo tipagem de dados com atributo '.dtype':", '\n',series_idades.dtype, '\n') # mesmo resultado com atributo '.dtypes'

print("Exibindo se a serie esta vazia com atributo '.empty':", '\n',series_idades.empty, '\n')

print("Exibindo se a serie possuí valores nulos com atributo '.hasnans':", '\n',series_idades.hasnans, '\n')

# Métodos

print("Exibindo valor máximo da série com método '.max()':", '\n', series_idades.max(), '\n')

print("Exibindo valor mínimo da série com método '.min()':", '\n', series_idades.min(), '\n')

print("Exibindo valor médio da série com método '.mean()':", '\n', series_idades.mean(), '\n')

print("Exibindo valor mediano da série com método '.median()':", '\n', series_idades.median(), '\n')

print("Exibindo valor ou valores moda da série com método '.mode()':", '\n', series_idades.mode(), '\n')

print("Exibindo estatísticas com método '.describe()':", '\n', series_idades.describe(), '\n')


### Operando com Series:

# Series x Series (Vetorial)

numeros_pares = pd.Series([*range(0,10,2)]) + 2 # Inteiros
numeros_impares = pd.Series([*range(0,10,2)]) + 1 # Inteiros
letras_a = pd.Series(indexs[:5]) # String
letras_b = pd.Series(indexs[5:10]) # String

soma = numeros_pares + numeros_impares
print(soma, '\n')

subtracao = numeros_pares - numeros_impares
print(subtracao, '\n')

concat = letras_a + letras_b
print(concat, '\n')

# Escalar

numeros_1a5 = pd.Series([*range(5)]) + 1 # Inteiros

mult_2 = numeros_1a5 * 2
print(mult_2, '\n')

mais_3 = numeros_1a5 + 3
print(mais_3, '\n')

quadrados = numeros_1a5 ** 2
print(quadrados, '\n')

cubos = numeros_1a5 ** 3
print(cubos, '\n')

# Lógicos

numeros_1a6 = pd.Series([*range(6)], index=[*range(1,7)]) + 1 # Inteiros


maiores_de_4 = numeros_1a6 > 4
print(maiores_de_4, '\n')

entre_2_4 = (numeros_1a6 >= 2) * (numeros_1a6 <= 4)
# entre_2_4 = (numeros_1a6 >= 2) & (numeros_1a6 <= 4)
print(entre_2_4, '\n')

ou_3_6 = (numeros_1a6 == 3) + (numeros_1a6 == 6)
# ou_3_6 = (numeros_1a6 == 3) | (numeros_1a6 == 6)
print(ou_3_6, '\n')
