==> 03_dataframes.ipynb <==
## DataFrames no Pandas

### Características:

- Conjunto de Valores bidimensional
- Agrupamento de Series Pandas
- Ordenado - A ordem defini o índices dos elementos
- Comportamento parecido com Tabelas
- Estrutura mais usada em manipulação de dados

# Importando bibliotécas

import pandas as pd 

# CRIANÇÃO

df_pandas = pd.DataFrame() #Explícta apenas

idades = [
    32, 38, 30, 30, 31,
    35, 25, 29, 31, 37,
    27, 23, 36, 33, 39
]

letras = [
    'a', 'b', 'c', 'd', 'e',
    'f', 'g', 'h', 'i', 'j',
    'k', 'l', 'm', 'n', 'o'
]

series_idades = pd.Series(idades)
series_letras = pd.Series(letras)

# Criando e atribuindo dados no dataframe

df_pandas['idades'] = series_idades
df_pandas['letras'] = series_letras

print(df_pandas)

### Navegando nos dados:

- `dataframe[ i ]` - Acessa Campos/Colunas a partir do nome do campo 
- `dataframe.loc[ i ]` - Acessa Registros/Linhas a partir do índice
- `dataframe.iloc[ i ]` - Acessa Registros/Linhas a partir da Posição

# dataframe[ i ] 

navega_df = df_pandas['letras']
print('df na coluna de nome "letras":', '\n', navega_df, '\n') 

navega_df = df_pandas['idades']
print('df na coluna de nome "idades":', '\n', navega_df, '\n') 


# dataframe.loc[ i ] 

indexs = [
    'i', 'ii', 'iii', 'iv', 'v',
    'vi', 'vii', 'viii', 'ix', 'x',
    'xi', 'xii', 'xiii', 'xiv', 'xv'
]

df_pandas.index = indexs # Alterando index
print(df_pandas, '\n')

navega_df = df_pandas.loc['xii']
print('df na linha de índice "xii":', '\n', navega_df) # Acessa índice da linha


# dataframe.iloc[ i ] 

navega_df = df_pandas.iloc[0]
print('df na linha da primeira posição:', '\n', navega_df) # Acessa posição da linha

navega_df = df_pandas.iloc[-2]
print('\n', 'df na linha da penultima posição:', '\n', navega_df)

### Outras formas úteis de navegação:

- `dataframe.loc[ i, "coluna" ]` - Acessa um valor específico a partir do índice da linha e nome da coluna 
- `dataframe.iloc[ i, j ]` - Acessa um valor específico pela posição numérica da linha e da coluna
- `dataframe.loc[ i : j ]` - Slicer de linhas baseado no índice
- `dataframe.iloc[ i : j ]` - Slicer de linhas baseado na posição


# dataframe.loc[ i, "coluna" ]

navega_valor = df_pandas.loc['v', 'letras']
print("valor no índice 'v' e na coluna 'letras':", '\n', navega_valor)

navega_valor = df_pandas.loc['x', 'letras']
print("valor no índice 'x' e na coluna 'letras':", '\n', navega_valor)

# dataframe.iloc[ i, j ]

navega_valor = df_pandas.iloc[5, 1]
print("valor na linha posição 5 e na coluna de posição 1:", '\n', navega_valor)

navega_valor = df_pandas.iloc[-1, 0]
print("valor na última posição e na coluna de posição 0", '\n', navega_valor)


# dataframe.loc[ i : j ]

navega_slice = df_pandas.loc['viii': 'xiii']
print("linhas do índice 'viii' ao índice 'xiii':", '\n', navega_slice, '\n')

navega_slice = df_pandas.loc['x':'xv']
print("linhas do índice 'x' ao índice 'xv':", '\n', navega_slice, '\n')

navega_slice = df_pandas.loc['iii':'i':-1]
print("linhas do índice 'iii' ao índice 'i':", '\n', navega_slice, '\n') # Reverse

navega_slice = df_pandas.loc['i':'ii':,'letras']
print("linhas do índice 'i' ao índice 'ii' na coluna 'letras':", '\n', navega_slice, '\n')



# dataframe.iloc[ i : j ]

navega_slice = df_pandas.iloc[7: 10]
print("linhas na posição 8 até posição 10:", '\n', navega_slice, '\n')

navega_slice = df_pandas.iloc[2:8]
print("linhas na posição 3 até posição 8:", '\n', navega_slice, '\n')

navega_slice = df_pandas.iloc[0:7:2]
print("linhas na posição 0 até posição 6 pulando um:", '\n', navega_slice, '\n') # Jump

navega_slice = df_pandas.iloc[:, 0:2]
print("todas as linhas nas colunas de posição 0 até posição 1:", '\n', navega_slice, '\n')


### Colunas:

# Selecinando Colunas

df_pandas['letras'] # retorna uma series

df_pandas[['letras']] # retorna um df de uma coluna

df_pandas[['letras', 'idades']] # retorna um df de mais de uma coluna


# Renomeando Colunas

series_numeros = pd.Series([*range(15)], index= indexs) + 1
df_pandas['numeros'] = series_numeros

df_pandas = df_pandas.rename(columns={"numeros": "a", "letras": "b"}) # recebe um dicionário contendo antes/depois para renomear

df_pandas.rename(columns={"idades": "c"}, inplace=True) # inplace anula necessidade de reatribuição

df_pandas

# Ordenando por ordem alfabética

colunas = df_pandas.columns.to_list()
colunas.sort()

df_pandas[colunas]

### Funções úteis:

# Atributos

print("Exibindo a quantidade de linhas e colunas com atributo '.shape':", '\n', df_pandas.shape, '\n')

print("Exibindo nomes de colunas em lista com atributo '.columns':", '\n',df_pandas.columns, '\n') 

print("Exibindo labels dos índices com atributo '.index':", '\n',df_pandas.index, '\n')

print("Exibindo tipagem de dados por coluna com atributo '.dtypes':", '\n',df_pandas.dtypes, '\n')

# Métodos

print("Exibindo primeiras 3 linhas com método '.head(3)':", '\n', df_pandas.head(3), '\n') # 5 linhas por defaut caso não informado 

print("Exibindo últimas 3 linhas com método '.tail(3)':", '\n', df_pandas.tail(3), '\n') # 5 linhas por defaut caso não informado

print("Exibindo 3 linhas escolhidas aleatóriamente com método '.sample(3)':", '\n', df_pandas.sample(3), '\n') # 1 linha por defaut caso não informado

print("Exibindo informações gerais com método '.info()':", '\n', df_pandas.info(), '\n') # memory usage = 'deep' motras consumo real de memoria
