==> 04_arquivos.ipynb <==
## Manipulando arquivos com Pandas

### Importando e exportando arquivos de diferentes extenções

# Importando bilibotécas

import pandas as pd 


# Importando arquivo CSV da pasta 'data'

df = pd.read_csv("data/clientes.csv") 
df


# Exportando arquivo CSV

df.to_csv("data/clientes.csv", index=False) # index=False ignora criação de índice

# Exportando arquivos de outros tipos

df.to_parquet("data/clientes.parquet", index=False)
df.to_excel("data/clientes.xlsx", index=False)

# Importando arquivo Parquet

df_2 = pd.read_parquet("data/clientes.parquet") # necessário bibliotéca de apoio pyarrow
df_2

# Importando arquivo Excel

df_3 = pd.read_excel("data/clientes.xlsx") # necessário bibliotéca de apoio openpyxl
df_3

# Ajustando separador para leitura

texto_csv = '''
sml; 32
jnf; 31
eve; 12
kev; 5
'''
with open("data/texto.csv", 'w') as file:
    file.write(texto_csv)

df_csv = pd.read_csv("data/texto.csv", sep=';') # estipula ';' como separador
df_csv

# Importando Área de Transferência

texto_csv = pd.DataFrame(df_csv).to_clipboard() # move arquivo para área de transferência

df_clip = pd.read_clipboard(sep=';') # abre arquivo da área de transferência
df_clip

# Importando HTML


url = "https://pt.wikipedia.org/wiki/Unidades_federativas_do_Brasil"

df_uf = pd.read_html(url) # retorna uma listas de tabelas encontradas na url

df_uf = df_uf[1]

df_uf.to_csv("data/UFs_Brasil.csv", sep= ';')
