==> 05_filtragem.ipynb <==
## Filtragem de dados no Pandas

### Operação muito importante em dados onde os dados de um df são filtrados a partir de uma operação lógica

# Importando bibliotécas

import pandas as pd

# Filtrando base de teste

df_filtro = pd.DataFrame(
    {
        "nome": ['Smaley', 'Jeniffer', 'Evelyn', 'Kevin'],
        "idade": [32, 31, 12, 5],
        "cidades_n": ['SGN', 'NLP', 'RJN', 'RJN']
    }
)

# Filtrando dataframe onde idade maior que 18 anos

filtro = df_filtro["idade"] > 18

df_filtro[filtro].head()

# Filtrando base real

df = pd.read_csv("../data/transacoes.csv")

# Filtrando transações cujo a quantidade de pontos é maior que 50

filtro = df["qtdePontos"] >= 50

df[filtro].head()

### Filtrando com base em mais de uma condição

# Filtrando pontos > 50 e < 100 usando operador "&"

filtro = (df["qtdePontos"] >= 50) & (df["qtdePontos"] <= 100) 
# filtro = (df["qtdePontos"] >= 50) * (df["qtdePontos"] <= 100) 

df[filtro].head()

# Filtrando pontos = 1 ou pontos = 100 usando operador "|"

filtro = (df["qtdePontos"] >= 50) | (df["qtdePontos"] <= 100)
# filtro = (df["qtdePontos"] >= 50) + (df["qtdePontos"] <= 100)

df[filtro].head()

# Filtrando pontos > 0 e < 50, e data >=2025

filtro = (df["qtdePontos"] > 0) & (df["qtdePontos"] < 50) & (df["dtCriacao"] >= '2025-01-01')
# filtro = (df["qtdePontos"] > 0) * (df["qtdePontos"] < 50) * (df["dtCriacao"] >= '2025-01-01')

df[filtro].head()

#Filtrando pontos > 0 e < 50, ou data >=2025

filtro = (df["qtdePontos"] > 0) & (df["qtdePontos"] < 50) | (df["dtCriacao"] >= '2025-01-01')
# filtro = (df["qtdePontos"] > 0) * (df["qtdePontos"] < 50) + (df["dtCriacao"] >= '2025-01-01')

df[filtro].head()

### Filtrando com uso de Métodos


# Usando método '.isin' e passando uma lista

df_novo =pd.read_csv("../data/transacao_produto.csv")

# filtro = (df_novo['idProduto'] == 1) | (df_novo['idProduto'] == 5)
filtro = df_novo['idProduto'].isin([1,5])

df_novo[filtro].head()


# Usando método '.notna', filtrando não vazias

df_novo2 =pd.read_csv("../data/clientes.csv")

# filtro = ~df_novo2['dtCriacao'].isna()
filtro = df_novo2['dtCriacao'].notna()

df_novo2[filtro].head()
