==> 06_transformacoes.ipynb <==
## Transformações de dados no Pandas

### Possibilidades:

- Criação de colunas a partir de operações com colunas
- Tipagem - convertendo tipo de dados de uma coluna
- Subistituição - Substituindo dados
- Trabalhando com Data

# Importando bibliotécas

import pandas as pd


# Operando e reatribuindo transformação

df = pd.read_csv("../data/clientes.csv")

df['qtdePontos'] = df['qtdePontos'] * 10

df.head()

# Operando e criando nova coluna

df['email_ou_twitch'] = df['flEmail'] + df['flTwitch']

df['email_e_twitch'] = df['flEmail'] * df['flTwitch']

df.head()

# Operando e criando nova coluna a partir varias colunas

df['qtdeSocial'] = df['flEmail'] + df['flTwitch'] + df['flYouTube'] + df['flBlueSky'] + df['flInstagram']

df['todasSocial'] = df['flEmail'] * df['flTwitch'] * df['flYouTube'] * df['flBlueSky'] * df['flInstagram']

df.head()

# Operando a partir de calculos envolvendo outras bibliotécas

import numpy as np

df['logPontos'] = np.log(df['qtdePontos'] + 1) # Logaritmo
df['logPontos'].describe()

# Convertendo tipagem de Series

print(df['qtdePontos'].dtypes)


df['qtdePontos'] = df['qtdePontos'].astype(float)

print("Tipagem transformada: ",df['qtdePontos'].dtypes)


# Substituindo valores 

df['dtCriacao'] = df['dtCriacao'].replace({ # recebe um dicionário
    "0000-00-00 00:00:00.000": "2024-02-01 09:00:00.000",
    })

# Convertendo tipagem para Data

df['dtCriacao'] = pd.to_datetime(df['dtCriacao'])

print(df['dtCriacao'].dtypes)

# Algumas funções de transformação de data - atributo '.dt'

filtro = df['dtCriacao'].notna()

df = df[filtro]

print(df['dtCriacao'].dt.date.sample(),'\n') # Somente data

print(df['dtCriacao'].dt.year.sample(),'\n') # Ano

print(df['dtCriacao'].dt.month.sample(),'\n') # Mês

print(df['dtCriacao'].dt.day.sample(),'\n') # Dia

print(df['dtCriacao'].dt.day_name().sample(),'\n') # Nome dia

print(df['dtCriacao'].dt.month_name().sample(),'\n') # Nome Mês
