==> 07_ordenacao.ipynb <==
## Ordenação de dados no Pandas

### Ordenando Series e Dataframes

# Importando bibliotécas

import pandas as pd


# Ordenando Series

clientes = pd.read_csv("../data/clientes.csv")

clientes['qtdePontos'].sort_values() # ordena valores de uma Series

# Ordenando Dataframes a partir de uma coluna

top_5 = (clientes.sort_values(by='qtdePontos', ascending=False)
                 .head(5)['idCliente'])
print("Top 5 Ids em relação a quantidade de pontos: ",'\n',top_5)

# Ordenando Dataframes a partir de mais de uma coluna

df_exemplo = {
    'nome': ['Smaley','Jeniffer', 'Evelyn', 'Kevin'],
    'idade': [32, 31, 12, 5],
    'pontos': [100, 100, 90, 80]
}

df_exemplo = pd.DataFrame(df_exemplo)

# Ordenando df por coluna pontos, desempate - coluna idade

desempate = (df_exemplo
                .sort_values(by=['pontos', 'idade'] , ascending=False)
                .head(4)['nome'])

print(desempate, '\n')

# Ordenando df por coluna pontos, desempate - coluna MENOR idade

desempate_x = (df_exemplo
                .sort_values(by=['pontos', 'idade'] , ascending=[False, True])
                .head(4)['nome'])

print(desempate_x, '\n')

# Ordenando df por coluna pontos, desempate - coluna MENOR índice

desempate_id = (df_exemplo
             .reset_index()
             .sort_values(by=['pontos', 'index'] , ascending=[False, True])
             .head(4)['nome'])

print(desempate_id)
