==> 08_na.ipynb <==
## Tratando dados ausentes no Pandas

### Usando métodos .dropna() e .fillna() para excluir e/ou preencher valores ausentes

# Importando bibliotécas

import pandas as pd

# dropna

df = pd.DataFrame(
    {
        'nome': [None, 'Jeniffer', 'Evelyn', 'Kevin', None],
        'idade': [None, None, 12, 5, None],
        'altura': [1.70, 1.75, None, 1.25, None]
    }
)

print(df, '\n')

# Excluindo linhas que apresentem qualquer uma coluna com valor nulo ou ausente

df_na_total = df.dropna()

print(".dropna(how='any'):\n",df_na_total, '\n')

# Excluindo linhas que apresentam valores nulos ou ausentes em todas as colunas

df_na_parcial = df.dropna(how='all')

print(".dropna(how='all'):\n",df_na_parcial, '\n')

# Excluindo linhas que apresentam somente valores nulos ou ausentes na coluna 'idade'

df_na_idade = df.dropna(how='all', subset=['idade'])
print(".dropna(how='all', subset=['idade']): \n",df_na_idade, '\n')

# Excluindo linhas que apresentam qualquer uma coluna com valores nulos ou ausentes entre as colunas 'idade' e 'nome'

df_na_idade_nome = df.dropna(how='any', subset=['nome', 'idade'])
print(".dropna(how='any', subset=['nome', 'idade']:\n",df_na_idade_nome, '\n')

# fillna

# Preenchendo valores nulos ou ausentes com '0' na coluna 'idade'

df_fill = df.copy()

df_fill['idade'] = df_fill['idade'].fillna(0)

print(".fillna(0):\n",df_fill, '\n') 

# Preenchendo valores nulos ou ausentes com '-' em qualquer coluna

df_fill = df_fill.fillna('-')

print(".fillna('-'):\n",df_fill, '\n')

# Preenchendo valores nulos ou ausentes passando instruções num dicionário

df_fill2= df.copy()

df_fill2 = df_fill2.fillna({'nome': '-', 'idade': 0})

print(".fillna({'nome': '-', 'idade': 0}):\n",df_fill2, '\n')


# Preenchendo valores nulos ou ausentes passando média de valores

df_fill3 = df.copy()

medias = df[['idade', 'altura']].mean()

df_fill3 = df_fill3.fillna(medias)

print("fillna(medias):\n",df_fill3, '\n')
