==> 09_duplicates.ipynb <==
## Tratando dados duplicados no Pandas

### Usando método .drop_duplicates para excluir valores duplicados

# Importando bibliotécas

import pandas as pd

# drop_duplicates

df = pd.DataFrame(
    {
        'nome': ['Smaley', 'Jeniffer', 'Jeniffer', 'Evelyn', 'Evelyn', 'Smaley', 'Kevin'],
        'sobrenome': ['Ferreira', 'Amorim', 'Marques', 'Marques','Marques', 'Marques', 'Marques'],
        'altura': [1.70, 1.65, 1.65, 1.58, 1.58, 1.71, 1.25]
    }
)

# Excluindo linhas que apresentem valores duplicados em todas as colunas

df_dropd_all = df.copy()

df_dropd_all = df_dropd_all.drop_duplicates()

print(".drop_duplicates():\n",df_dropd_all,'\n')

# Excluindo linhas que apresentem valores duplicados na coluna nome

df_dropd_nome = df.copy()

df_dropd_nome = df_dropd_nome.drop_duplicates('nome')

print(".drop_duplicates('nome'):\n",df_dropd_nome,'\n')

# Excluindo linhas que apresentem valores duplicados na coluna nome, mantendo ultima ocorrência

df_dropd_nome = df.copy()

df_dropd_nome = df_dropd_nome.drop_duplicates('nome', keep='last')

print(".drop_duplicates('nome', keep='last'):\n",df_dropd_nome,'\n')

# Excluindo linhas que apresentem valores duplicados nas colunas 'nome' e 'altura', mantendo ultima ocorrência de ambos

df_dropd_nome_alt = df.copy()

df_dropd_nome_alt = df_dropd_nome_alt.drop_duplicates(subset=['nome', 'altura'], keep='last')

print(".drop_duplicates(subset=['nome', 'altura'], keep='last'):\n",df_dropd_nome_alt,'\n')
