==> 10_apply.ipynb <==
## Aplicando funções personalizadas no Pandas

### Usando método .apply

# Importando bibliotécas

import pandas as pd

# apply

df = pd.read_csv("../data/clientes.csv")

# Aplicando função personalizada na coluna 'idCliente'

def get_lst_id(x): # Declarando função que recebe um valor da 'idCliente' e separa mantendo última parte
    return x.split('-')[-1]

# Aplicando função na coluna

df['IdNovo'] = df['idCliente'].apply(get_lst_id)

df.head()

uf = pd.read_csv("../Módulo_02/UFs_Brasil.csv", sep=";")

# Aplicando função personalizada em varias colunas

def str_to_float(x:str): # Declarando função que trata str e transforma em float
    x = float(x.replace(" ","")
               .replace(",", ".")
               .replace("\xa0", ""))
    return x

colunas = ['Área (km²)', 'População (Censo 2022)', 'PIB (2015)', 'PIB per capita (R$) (2015)']

# Aplicando função com for

for i in colunas:
    uf[i] = uf[i].apply(str_to_float)

uf.dtypes

# Aplicando função para criação de uma coluna a partir de valores de outra

def uf_to_regiao(uf): # Declarando função que avalia estado e retorna região

    if uf in ["Distrito Federal", "Goiás", "Mato Grosso", "Mato Grosso do Sul"]:
        return "Centro-Oeste"
    elif uf in ["Alagoas","Bahia", "Ceará", "Maranhão", "Paraíba", "Pernambuco", "Piauí", "Rio Grande do Norte", "Sergipe"]:
        return "Nordeste"
    elif uf in ["Acre", "Amapá", "Amazonas", "Pará", "Rondônia", "Roraima", "Tocantins"]:
       return "Norte"
    elif uf in ["Espírito Santo","Minas Gerais", "Rio de Janeiro", "São Paulo"]:
        return "Sudeste"
    elif uf in ["Paraná", "Rio Grande do Sul", "Santa Catarina"]:
        return "Sul"
    
# Aplicando função na coluna e atraibuindo em nova coluna

uf["Região"] = uf["Unidade federativa"].apply(uf_to_regiao)

uf.head()

# Aplicando função personalizada em linhas

def pib_per_capita_manual(x): # Declarando função que calcula a partir de duas colunas
    return x['PIB (2015)'] / x['População (Censo 2022)']

# Aplicando função com argumento axis=1 para aplicação linha a linha

uf['PIB per capita manual'] = uf.apply(pib_per_capita_manual, axis=1)

uf.head()
