==> 11_agregacoes.ipynb <==
## Agregações de dados no Pandas

### Agrupando e sumarizando Series e DataFrames

# Importando bibliotécas

import pandas as pd

# Agregações gerais

clientes = pd.read_csv("data/clientes.csv")

# Agregando valores de uma coluna 

print("Soma de todos os valores da colunha 'flTwitch': ", clientes['flTwitch'].sum(), '\n')
print("Média de todos os valores da colunha 'flTwitch': ", clientes['flTwitch'].mean(), '\n')

# Agregando valores de mais de uma coluna

print("Estatísticas dos valores das colunas 'flTwitch', 'flYouTube':\n",clientes[['flTwitch', 'flYouTube']].describe(), '\n')

sociais = ['flEmail', 'flTwitch', 'flYouTube', 'flBlueSky', 'flInstagram']

print("Estatísticas dos valores de todas as colunas referentes a Rede Social:\n",clientes[sociais].describe(), '\n')

# Agregando valores das colunas numéricas

num_columns = clientes.dtypes[~(clientes.dtypes == 'object')].index.tolist()

clientes[num_columns].describe()

# groupby

transacoes = pd.read_csv("data/transacoes.csv")

# Agregando a partir de uma coluna

transacoes.groupby(by=['idCliente']).count() # Retorna um df

# Agregando a partir de uma coluna por uma coluna

transacoes.groupby(by=['idCliente'])[['idTransacao']].count() # Retorna um df

# transacoes.groupby(by=['idCliente'])['idTransacao'].count() # Retorna uma series

# Modificando índice

transacoes.groupby(by=['idCliente'], as_index=False)[['idTransacao']].count()

# aggregate

# Agregando multiplas estatísticas de uma vez

summary = (transacoes.groupby(by=["idCliente"], as_index=False)
                       .aggregate({'idTransacao': ['count'], # Recebe um dicionário
                                   'qtdePontos': ['sum', 'mean']})) 

print(summary.head(), '\n')

# Acessando colunas multinível

print(summary['qtdePontos'].head(), '\n') # Acessando colunas pais

print(summary['qtdePontos']['sum'].head()) # Acessando colunas filhas

# Renomeando colunas multinível

summary.columns = ['idCliente', 'qtdeTransacoes', 'TotalPontos', 'MédiaPontos']

summary.head()

# Agregando função personalizada

def diff_dias (x: pd.Series):
    date = pd.to_datetime(x)
    return (date.max() - date.min()).days

summary_novo = (transacoes.groupby(by=["idCliente"], as_index=False)
                          .agg({'idTransacao': ['count'], # agg = aggregate
                                      'qtdePontos': ['sum', 'mean'],
                                      'dtCriacao': [diff_dias]})) 

summary_novo
