==> 12_mesclagens.ipynb <==
## Mesclagem de dados no Pandas

### Mesclando dados de diferentes DataFrames

# Importando bibliotécas

import pandas as pd

# merge

clientes = pd.read_csv("../data/clientes.csv")
transacoes = pd.read_csv("../data/transacoes.csv")

transacoes.merge(
    right=clientes, # Tabela a mesclar
    how='left', # Modo de mesclagem
    on=['idCliente'], # Coluna em comum
    suffixes=['Transacao', 'Clientes'] # Sufixo a anexar em caso de redundância
) 

# Mesclando a partir de mais de uma coluna 

transacoes.merge(
    right=clientes, # Tabela a mesclar
    how='left', # Modo de mesclagem
    on = ['idCliente', 'dtCriacao'], # Colunas em comum
    suffixes=['Transacao', 'Clientes'] # Sufixo a anexar em caso de redundância
)

# Mesclando a partir de colunas de nomes diferentes

clientes.rename(columns={'idCliente': 'Cliente'}, inplace=True)

transacoes.merge(
    right=clientes, # Tabela a mesclar
    how='left', # Modo de mesclagem
    left_on=['idCliente'], # Coluna da tabela referente
    right_on=['Cliente'], # Coluna da tabela busca
    suffixes=['Transacao', 'Clientes'] # Sufixo a anexar em caso de redundância
)
