==> 13_concatenacoes.ipynb <==
## Concatenação de dados no Pandas

### Concatenando dados de diferentes formas

# Importando bibliotécas

import pandas as pd

# concat

df1 = pd.DataFrame(
    {
        "nome": ['Smaley', 'Jeniffer', 'Evelyn', 'Kevin'],
        "idade": [32, 31, 12, 5],
        "cidades_n": ['SGN', 'NLP', 'RJN', 'RJN']
    }
)

df2 = pd.DataFrame(
    {
        "nome": ['Smaley', 'Jeniffer', 'Evelyn', 'Kevin'],
        "cidades_n": ['SGN', 'NLP', 'RJN', 'RJN']
    }
)

# Empilhando Dataframes

print(pd.concat([df1, df2]),'\n')

# Empilhando Dataframes e rearranjando índices

pd.concat([df1, df2], ignore_index=True)

# Concatenando Dataframes (lado a lado)

df3 = pd.DataFrame(
    {'idades': [32,31,12,5,8]}
)

print(pd.concat([df2, df3], axis=1))

# Concatenando Dataframes a partir do índice

df3 = df3.sort_values(by='idades').reset_index(drop=True)

pd.concat([df2, df3], axis=1)
