==> 14_db.ipynb <==
## Conexão com Banco de Dados no Pandas

### Conectando e manipulando dados de bancos de dados

# Importando bibliotécas

import pandas as pd
import sqlalchemy

# Criando engine de conexão / string de conexão

engine = sqlalchemy.create_engine("sqlite:///../data/olist.db")

# Carregando tabela a partir da engine

clientes = pd.read_sql_table(table_name="tb_customers", con=engine)

print(clientes.shape)
clientes.head()

# Carregando query SQL

query = "SELECT * FROM tb_customers LIMIT 100"

df = pd.read_sql_query(query, engine)

print(df.shape)
df.head()

# Trabalhando query salva externamente

with open("etl.sql") as open_file:
    query = open_file.read()

print(query)

df = pd.read_sql_query(query, engine)

df.head()

# Salvando tabela no banco de dados

inspector = sqlalchemy.inspect(engine) # Instanciando o inpencionador de db

df.to_sql("tb_teste_table", con= engine, index=False, if_exists='replace')

print(inspector.get_table_names())
