==> 15_stack.ipynb <==
## Invertendo linhas e colunas no Pandas

### *.stack, .unstack e .pivot* 

# Importando bibliotécas

import pandas as pd

# stack

df = pd.read_csv("data/clientes.csv", usecols=['idCliente', 'flEmail', 'flTwitch', 'flYouTube', 'flBlueSky', 'flInstagram'])

# Empilhando colunas em linhas

df_stack = (df.set_index('idCliente')
              .stack()
              .reset_index())

df_stack.columns = ['idCliente', 'RedeSocial', 'Valor']

df_stack.head()

# unstack

# Desempilhando linhas em colunas

df_unstack = (df_stack.set_index(['idCliente', 'RedeSocial'])
                      .unstack()
                      .reset_index())

# Tratando Multindex das colunas

redes = df_unstack.columns.droplevel().tolist()[1:]
df_unstack.columns = ['idCliente'] + redes

df_unstack.head()

# pivot_table

# Pivoetando tabela com agregação

df_pivot = (df_stack.pivot_table(values='Valor',
                                 index=['idCliente'],
                                 columns='RedeSocial',
                                 aggfunc='sum')
                    .reset_index())

df_pivot.head()
